pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
include = ["edabot"]

[project]
name = "edabot"
version = "0.1.0"
description = "An AI assistant for exploratory data analysis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.49.0",
    "chatlas>=0.4.0",
    "ipykernel>=6.29.5",
    "openai>=1.65.4",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "polars>=1.25.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
    "syrupy>=4.9.1",
]

[tool.ruff.lint]
exclude = ["docs", ".venv", "tests/*"]
